Factor out the common implementation of height & width properties.
