+++ src/plugins/drawing/objects/image.tsx
@@ -2,13 +2,12 @@ import { Instance, SnapshotIn, types } from "mobx-state-tree";
 import React from "react";
 import { observer } from "mobx-react";
 import { gImageMap } from "../../../models/image-map";
-import { DrawingObject, DrawingObjectSnapshot, DrawingTool,
-  IDrawingComponentProps, IDrawingLayer, ObjectTypeIconViewBox, typeField } from "./drawing-object";
-import { BoundingBoxSides, Point } from "../model/drawing-basic-types";
+import { DrawingObjectSnapshot, DrawingTool,
+  IDrawingComponentProps, IDrawingLayer, ObjectTypeIconViewBox, SizedObject, typeField } from "./drawing-object";
 import placeholderImage from "../../../assets/image_placeholder.png";
 import ImageToolIcon from "../../../clue/assets/icons/image-tool.svg";
 
-export const ImageObject = DrawingObject.named("ImageObject")
+export const ImageObject = SizedObject.named("ImageObject")
   .props({
     type: typeField("image"),
 
@@ -21,22 +20,8 @@ export const ImageObject = DrawingObject.named("ImageObject")
     // into a image tile and then that image dragged to the drawing tile. In this case we want
     // to preserve the filename of the source image.
     filename: types.maybe(types.string),
-    width: types.number,
-    height: types.number
   })
-  .volatile(self => ({
-    dragWidth: undefined as number | undefined,
-    dragHeight: undefined as number | undefined
-  }))
   .views(self => ({
-    get boundingBox() {
-      const {x, y} = self.position;
-      const width = self.dragWidth ?? self.width;
-      const height = self.dragHeight ?? self.height;
-      const nw: Point = {x, y};
-      const se: Point = {x: x + width, y: y + height};
-      return {nw, se};
-    },
     get label() {
       return "Image";
     },
@@ -63,18 +48,6 @@ export const ImageObject = DrawingObject.named("ImageObject")
       self.filename = filename;
     },
 
-    setDragBounds(deltas: BoundingBoxSides) {
-      self.dragX = self.x + deltas.left;
-      self.dragY = self.y + deltas.top;
-      self.dragWidth  = self.width  + deltas.right - deltas.left;
-      self.dragHeight = self.height + deltas.bottom - deltas.top;
-    },
-    resizeObject() {
-      self.repositionObject();
-      self.width = self.dragWidth ?? self.width;
-      self.height = self.dragHeight ?? self.height;
-      self.dragWidth = self.dragHeight = undefined;
-    }
   }));
 export interface ImageObjectType extends Instance<typeof ImageObject> {}
 export interface ImageObjectSnapshot extends SnapshotIn<typeof ImageObject> {}
