+++ src/plugins/drawing/objects/rectangle.tsx
@@ -2,37 +2,17 @@ import { observer } from "mobx-react";
 import { Instance, SnapshotIn, types, getSnapshot } from "mobx-state-tree";
 import React from "react";
 import { computeStrokeDashArray, DrawingTool, FilledObject, IDrawingComponentProps,
-  IDrawingLayer, ObjectTypeIconViewBox, StrokedObject, typeField } from "./drawing-object";
-import { BoundingBoxSides, Point } from "../model/drawing-basic-types";
+  IDrawingLayer, ObjectTypeIconViewBox, SizedObject, StrokedObject, typeField } from "./drawing-object";
+import { Point } from "../model/drawing-basic-types";
 import RectToolIcon from "../assets/rectangle-icon.svg";
 
-export const RectangleObject = types.compose("RectangleObject", StrokedObject, FilledObject)
+// Note: SizedObject must be listed last because it overrides the default implementation
+// of the boundingBox property.
+export const RectangleObject = types.compose("RectangleObject", StrokedObject, FilledObject, SizedObject)
   .props({
-    type: typeField("rectangle"),
-    width: types.number,
-    height: types.number,
+    type: typeField("rectangle")
   })
-  .volatile(self => ({
-    dragWidth: undefined as number | undefined,
-    dragHeight: undefined as number | undefined
-  }))
   .views(self => ({
-    get currentDims() {
-      const { width, height, dragWidth, dragHeight } = self;
-      return {
-        width: dragWidth ?? width,
-        height: dragHeight ?? height
-      };
-    }
-  }))
-  .views(self => ({
-    get boundingBox() {
-      const { x, y } = self.position;
-      const { width, height } = self.currentDims;
-      const nw: Point = {x, y};
-      const se: Point = {x: x + width, y: y + height};
-      return {nw, se};
-    },
     get label() {
       return self.width===self.height ? "Square" : "Rectangle";
     },
@@ -69,18 +49,6 @@ export const RectangleObject = types.compose("RectangleObject", StrokedObject, F
         self.y = y;
         self.width = self.height = squareSize;
       }
-    },
-    setDragBounds(deltas: BoundingBoxSides) {
-      self.dragX = self.x + deltas.left;
-      self.dragY = self.y + deltas.top;
-      self.dragWidth  = Math.max(self.width  + deltas.right - deltas.left, 1);
-      self.dragHeight = Math.max(self.height + deltas.bottom - deltas.top, 1);
-    },
-    resizeObject() {
-      self.repositionObject();
-      self.width = self.dragWidth ?? self.width;
-      self.height = self.dragHeight ?? self.height;
-      self.dragWidth = self.dragHeight = undefined;
     }
   }));
 export interface RectangleObjectType extends Instance<typeof RectangleObject> {}
